Use require-all for tests
